+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to **Traffic** is documented in this *changelog*.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **Traffic** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Zoom-in tooltip for all lists.
 ### Changed
 - Better selector layouts on mobile devices.
 ### Fixed